Move ChatClientFactory and Environment to separate files
