+++ Sources/StreamChat/ChatClient.swift
@@ -504,297 +504,3 @@ extension ClientError {
         }
     }
 }
-
-class ChatClientFactory {
-    let config: ChatClientConfig
-    let environment: ChatClient.Environment
-
-    init(config: ChatClientConfig, environment: ChatClient.Environment) {
-        self.config = config
-        self.environment = environment
-    }
-
-    func makeApiClientRequestEncoder() -> RequestEncoder {
-        environment.requestEncoderBuilder(config.baseURL.restAPIBaseURL, config.apiKey)
-    }
-
-    func makeWebSocketRequestEncoder() -> RequestEncoder {
-        environment.requestEncoderBuilder(config.baseURL.webSocketBaseURL, config.apiKey)
-    }
-
-    func makeApiClient(
-        encoder: RequestEncoder,
-        urlSessionConfiguration: URLSessionConfiguration
-    ) -> APIClient {
-        let decoder = environment.requestDecoderBuilder()
-        let attachmentUploader = config.customAttachmentUploader ?? StreamAttachmentUploader(
-            cdnClient: config.customCDNClient ?? StreamCDNClient(
-                encoder: encoder,
-                decoder: decoder,
-                sessionConfiguration: urlSessionConfiguration
-            )
-        )
-        let apiClient = environment.apiClientBuilder(
-            urlSessionConfiguration,
-            encoder,
-            decoder,
-            attachmentUploader
-        )
-        return apiClient
-    }
-
-    func makeWebSocketClient(
-        requestEncoder: RequestEncoder,
-        urlSessionConfiguration: URLSessionConfiguration,
-        eventNotificationCenter: EventNotificationCenter
-    ) -> WebSocketClient? {
-        environment.webSocketClientBuilder?(
-            urlSessionConfiguration,
-            requestEncoder,
-            EventDecoder(),
-            eventNotificationCenter
-        )
-    }
-
-    func makeDatabaseContainer() -> DatabaseContainer {
-        do {
-            if config.isLocalStorageEnabled {
-                guard let storeURL = config.localStorageFolderURL else {
-                    throw ClientError.MissingLocalStorageURL()
-                }
-
-                // Create the folder if needed
-                try FileManager.default.createDirectory(
-                    at: storeURL,
-                    withIntermediateDirectories: true,
-                    attributes: nil
-                )
-
-                let dbFileURL = storeURL.appendingPathComponent(config.apiKey.apiKeyString)
-                return environment.databaseContainerBuilder(
-                    .onDisk(databaseFileURL: dbFileURL),
-                    config.shouldFlushLocalStorageOnStart,
-                    config.isClientInActiveMode, // Only reset Ephemeral values in active mode
-                    config.localCaching,
-                    config.deletedMessagesVisibility,
-                    config.shouldShowShadowedMessages
-                )
-            }
-
-        } catch is ClientError.MissingLocalStorageURL {
-            log.assertionFailure("The URL provided in ChatClientConfig can't be `nil`. Falling back to the in-memory option.")
-
-        } catch {
-            log.error("Failed to initialize the local storage with error: \(error). Falling back to the in-memory option.")
-        }
-
-        return environment.databaseContainerBuilder(
-            .inMemory,
-            config.shouldFlushLocalStorageOnStart,
-            config.isClientInActiveMode, // Only reset Ephemeral values in active mode
-            config.localCaching,
-            config.deletedMessagesVisibility,
-            config.shouldShowShadowedMessages
-        )
-    }
-
-    func makeEventNotificationCenter(
-        databaseContainer: DatabaseContainer,
-        currentUserId: @escaping () -> UserId?
-    ) -> EventNotificationCenter {
-        let center = environment.notificationCenterBuilder(databaseContainer)
-        let middlewares: [EventMiddleware] = [
-            EventDataProcessorMiddleware(),
-            TypingStartCleanupMiddleware(
-                emitEvent: { [weak center] in center?.process($0) }
-            ),
-            ChannelReadUpdaterMiddleware(
-                newProcessedMessageIds: { [weak center] in center?.newMessageIds ?? [] }
-            ),
-            UserTypingStateUpdaterMiddleware(),
-            ChannelTruncatedEventMiddleware(),
-            MemberEventMiddleware(),
-            UserChannelBanEventsMiddleware(),
-            UserWatchingEventMiddleware(),
-            UserUpdateMiddleware(),
-            ChannelVisibilityEventMiddleware(),
-            EventDTOConverterMiddleware()
-        ]
-
-        center.add(middlewares: middlewares)
-
-        return center
-    }
-}
-
-extension ChatClient {
-    /// An object containing all dependencies of `Client`
-    struct Environment {
-        var apiClientBuilder: (
-            _ sessionConfiguration: URLSessionConfiguration,
-            _ requestEncoder: RequestEncoder,
-            _ requestDecoder: RequestDecoder,
-            _ attachmentUploader: AttachmentUploader
-        ) -> APIClient = {
-            APIClient(
-                sessionConfiguration: $0,
-                requestEncoder: $1,
-                requestDecoder: $2,
-                attachmentUploader: $3
-            )
-        }
-
-        var webSocketClientBuilder: ((
-            _ sessionConfiguration: URLSessionConfiguration,
-            _ requestEncoder: RequestEncoder,
-            _ eventDecoder: AnyEventDecoder,
-            _ notificationCenter: EventNotificationCenter
-        ) -> WebSocketClient)? = {
-            WebSocketClient(
-                sessionConfiguration: $0,
-                requestEncoder: $1,
-                eventDecoder: $2,
-                eventNotificationCenter: $3
-            )
-        }
-
-        var databaseContainerBuilder: (
-            _ kind: DatabaseContainer.Kind,
-            _ shouldFlushOnStart: Bool,
-            _ shouldResetEphemeralValuesOnStart: Bool,
-            _ localCachingSettings: ChatClientConfig.LocalCaching?,
-            _ deletedMessageVisibility: ChatClientConfig.DeletedMessageVisibility?,
-            _ shouldShowShadowedMessages: Bool?
-        ) -> DatabaseContainer = {
-            DatabaseContainer(
-                kind: $0,
-                shouldFlushOnStart: $1,
-                shouldResetEphemeralValuesOnStart: $2,
-                localCachingSettings: $3,
-                deletedMessagesVisibility: $4,
-                shouldShowShadowedMessages: $5
-            )
-        }
-
-        var extensionLifecycleBuilder = NotificationExtensionLifecycle.init
-
-        var requestEncoderBuilder: (_ baseURL: URL, _ apiKey: APIKey) -> RequestEncoder = DefaultRequestEncoder.init
-        var requestDecoderBuilder: () -> RequestDecoder = DefaultRequestDecoder.init
-
-        var eventDecoderBuilder: () -> EventDecoder = EventDecoder.init
-
-        var notificationCenterBuilder = EventNotificationCenter.init
-
-        var internetConnection: (_ center: NotificationCenter, _ monitor: InternetConnectionMonitor) -> InternetConnection = {
-            InternetConnection(notificationCenter: $0, monitor: $1)
-        }
-
-        var internetMonitor: InternetConnectionMonitor {
-            if let monitor = monitor {
-                return monitor
-            } else if #available(iOS 12, *) {
-                return InternetConnection.Monitor()
-            } else {
-                return InternetConnection.LegacyMonitor()
-            }
-        }
-
-        var monitor: InternetConnectionMonitor?
-
-        var connectionRepositoryBuilder = ConnectionRepository.init
-
-        var backgroundTaskSchedulerBuilder: () -> BackgroundTaskScheduler? = {
-            if Bundle.main.isAppExtension {
-                // No background task scheduler exists for app extensions.
-                return nil
-            } else {
-                #if os(iOS)
-                return IOSBackgroundTaskScheduler()
-                #else
-                // No need for background schedulers on macOS, app continues running when inactive.
-                return nil
-                #endif
-            }
-        }
-
-        var timerType: Timer.Type = DefaultTimer.self
-
-        var tokenExpirationRetryStrategy: RetryStrategy = DefaultRetryStrategy()
-
-        var connectionRecoveryHandlerBuilder: (
-            _ webSocketClient: WebSocketClient,
-            _ eventNotificationCenter: EventNotificationCenter,
-            _ syncRepository: SyncRepository,
-            _ extensionLifecycle: NotificationExtensionLifecycle,
-            _ backgroundTaskScheduler: BackgroundTaskScheduler?,
-            _ internetConnection: InternetConnection,
-            _ keepConnectionAliveInBackground: Bool
-        ) -> ConnectionRecoveryHandler = {
-            DefaultConnectionRecoveryHandler(
-                webSocketClient: $0,
-                eventNotificationCenter: $1,
-                syncRepository: $2,
-                extensionLifecycle: $3,
-                backgroundTaskScheduler: $4,
-                internetConnection: $5,
-                reconnectionStrategy: DefaultRetryStrategy(),
-                reconnectionTimerType: DefaultTimer.self,
-                keepConnectionAliveInBackground: $6
-            )
-        }
-
-        var authenticationRepositoryBuilder = AuthenticationRepository.init
-
-        var syncRepositoryBuilder: (
-            _ config: ChatClientConfig,
-            _ activeChannelControllers: ThreadSafeWeakCollection<ChatChannelController>,
-            _ activeChannelListControllers: ThreadSafeWeakCollection<ChatChannelListController>,
-            _ offlineRequestsRepository: OfflineRequestsRepository,
-            _ eventNotificationCenter: EventNotificationCenter,
-            _ database: DatabaseContainer,
-            _ apiClient: APIClient
-        ) -> SyncRepository = {
-            SyncRepository(
-                config: $0,
-                activeChannelControllers: $1,
-                activeChannelListControllers: $2,
-                offlineRequestsRepository: $3,
-                eventNotificationCenter: $4,
-                database: $5,
-                apiClient: $6
-            )
-        }
-
-        var callRepositoryBuilder: (
-            _ apiClient: APIClient
-        ) -> CallRepository = {
-            CallRepository(apiClient: $0)
-        }
-
-        var channelRepositoryBuilder: (
-            _ database: DatabaseContainer,
-            _ apiClient: APIClient
-        ) -> ChannelRepository = {
-            ChannelRepository(database: $0, apiClient: $1)
-        }
-
-        var messageRepositoryBuilder: (
-            _ database: DatabaseContainer,
-            _ apiClient: APIClient
-        ) -> MessageRepository = {
-            MessageRepository(database: $0, apiClient: $1)
-        }
-
-        var offlineRequestsRepositoryBuilder: (
-            _ messageRepository: MessageRepository,
-            _ database: DatabaseContainer,
-            _ apiClient: APIClient
-        ) -> OfflineRequestsRepository = {
-            OfflineRequestsRepository(
-                messageRepository: $0,
-                database: $1,
-                apiClient: $2
-            )
-        }
-    }
-}
